requirements.txt
pdf2docx>=0.5.6
python-docx>=1.0.0
docx2pdf>=0.1.8
Pillow>=10.0.0
pikepdf>=9.0.0
